레이저 쏘기
아직 흐접하다

+++ BirdShooter/Assets/Script/EnemyControl.cs
@@ -75,7 +75,7 @@ public class EnemyControl : MonoBehaviour {
         }
     }
 
-    void Damaged(int dam)
+    void Damaged(float dam)
     {
         mInfo.Health -= dam;
     }
